pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autograph"
version = "0.1.0"
description = "Dataset-agnostic knowledge graph builder with optional LLM and Neo4j integrations."
readme = "docs/goals.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.25.0,<1.0.0",
    "neo4j>=5.20,<6.0",
    "numpy>=1.26.0,<3.0.0",
    "openai>=2.26.0,<3.0.0",
    "openpyxl>=3.1.0,<4.0.0",
    "pandas>=2.2.0,<3.0.0",
    "pyarrow>=16.0.0,<25.0.0",
    "pydantic>=2.7.0,<3.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "tenacity>=8.2.3,<10.0.0",
    "tqdm>=4.66.0,<5.0.0",
]

authors = [
    { name = "AutoGraph Contributors" },
]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[project.urls]
Homepage = "https://github.com/cds0987/autograph"
Repository = "https://github.com/cds0987/autograph"
Issues = "https://github.com/cds0987/autograph/issues"

[tool.setuptools]
package-dir = { autograph = "." }
packages = [
    "autograph",
    "autograph.analyzers",
    "autograph.core",
    "autograph.embeddings",
    "autograph.graph_store",
    "autograph.loaders",
    "autograph.utils",
]
